Better events for force layout.
The force layout's listeners are now scoped using the "force" namespace,
allowing other event listeners to be registered. In addition, the force layout
now stops propagation of mouseup events when a node is moved; this allows a
mouseup listener on the node to fire only when it is not moved.