Update dependencies and use "devDependencies".
Using "devDependencies" prevents "vows" etc. from being installed if D3
is being installed as a dependency. I'm not sure whether "jsdom" counts
as a true dependency or not as we only really use it for the tests.

I've also added a "scripts": {"test": …} parameter, which specifies what
happens when you run "npm test"