Add PivotGraph (rollup) layout example.
This is ported from the Protovis implementation, pv.Layout.Rollup.
The only difference is that the rollup nodes are not instances of any
original nodes (Protovis used pv.extend).  Users are expected to use the
"x" and "y" properties of the rollup node, or alternatively the "rolled
up" nodes can be retrieved via the "nodes" property.