Simpler workaround for local d3.xhr.
If the error flag is set, the response attribute [1] should be the empty
string or null.  So even if a request fails and the response has been
partially loaded, this logic should still work.

This check will fail for local requests where the response entity body
is the empty string, but I don't think it's possible to detect this
situation in current browser implementations. jQuery has the same issue,
for example.

[1]: http://www.w3.org/TR/XMLHttpRequest/#the-response-attribute