Improve resampling and antemeridian cutting.
Lines going through poles are now detected, and multiple points are
inserted at the poles to aid resampling "along" a pole.

I've removed the distortion check for resampling, since this is no
longer necessary (this kind of distortion only really happens near
poles, which the cutting now takes care of).